added virtual destrutors to polymorphic base classes and reorganized a little
