+++ src/aldente.h
@@ -6,7 +6,7 @@
 #include <stdlib.h>
 #include <iostream>
 
-#include "main_scene.h"
+#include "scene/main_scene.h"
 #include "scene/scene_camera.h"
 
 /*
