+++ src/asset_loader.cpp
@@ -1,4 +1,4 @@
-#include "assetLoader.h"
+#include "asset_loader.h"
 #include "aldente.h"
 #include "util/colors.h"
 #include "shaders/shader_manager.h"
