+++ src/asset_loader.h
