+++ src/tile.h
@@ -2,14 +2,14 @@
 * Tile Class:
 *   This file holds the superclass of Tile and all of its subclasses. The superclass Tile holds basic
 *   varaibles and functions that all Tiles will require in order to function. These include, and are not
-*   limited to: width, height, x coord, z coord, Mesh, getter functions. Each actual Tile will then 
+*   limited to: width, height, x coord, z coord, Mesh, getter functions. Each actual Tile will then
 *   be a child of this superclass and implement the neccessary functionality for that specific tile
 *   (for example: rendering, collision detection, trigger functionality).
 */
 
 #pragma once
 
-#include "mesh.h"
+#include "model/mesh.h"
 #include "model/geometry_generator.h"
 #include "btBulletDynamicsCommon.h"
 
@@ -27,12 +27,12 @@ private:
 
 public:
 	Tile();
-	~Tile();
+	virtual ~Tile();
 
 	Mesh* getMesh() { return mesh; };
 	int getX() { return x; };
 	int getZ() { return z; };
-	virtual void update(Tile* hover) {};
+	virtual void update(Tile* hover) = 0;
 	btRigidBody* getRigid() { return rigidBody; };
 };
